requirements.txt
streamlit==1.53.0
google-genai>=0.2.0
python-dotenv>=1.0.0
